changed pin to string and balance to double
confirmed

+++ src/main/java/rocks/zipcode/atm/bank/Account.java
@@ -37,11 +37,11 @@ public abstract class Account {
         }
     }
 
-    public int getBalance() {
+    public Double getBalance() {
         return accountData.getBalance();
     }
 
-    private void updateBalance(int newBalance) {
+    private void updateBalance(Double newBalance) {
         accountData = new AccountData(accountData.getId(), accountData.getName(), accountData.getEmail(),
                 newBalance, accountData.getPin());
     }
